Refactor CommitAbility node into TryCommit
